==> Code/2 data_clean.ipynb <==
# <a id='toc1_'></a>[Data Cleaning and Processing](#toc0_)

# reset (removing all variables, functions, and other objects from memory)
%reset -f

# import all the necessary packages
import pandas as pd
import numpy as np
import warnings
import re



In this workbook we will clean and process the data for the project. We will also create a new dataset that will be used for the analysis. This dataset will be saved as a csv file and will be used in the analysis workbook. Specifically, we will do the following:

1. Load the data from the csv file created in the data loadings workbook
2. Clean the data by:
    - removing or looking at columns that are not needed
    - looking at missing values
    - looking at duplicates
    - looking at outliers
    - standardizing and checking data types
    - deal with text data
    - deal with categorical data (categories)
    - check for data consistency across columns
    - look at normalizing data (price, votes, etc. )
    - Looking at rows with missing values
    - check data balance

3. Create a new dataset
4. Save the new dataset

# <a id='toc2_'></a>[Data Cleaning Processes](#toc0_)

Applying cleaning process for each set of data. 

A set defines the least number reviews each reviewer and product has.

# load data - MAC OS
# amz_rev = pd.read_csv('/Users/pavansingh/Library/CloudStorage/GoogleDrive-pavansingho23@gmail.com/My Drive/Portfolio/Masters-Dissertation/Code/Data/all_revs_with_records_1.csv', low_memory=False)

# # load data - Set 1
amz_rev = pd.read_csv('/Users/pavansingh/Library/CloudStorage/GoogleDrive-pavansingho23@gmail.com/My Drive/Portfolio/Masters-Dissertation/Code/Data/set1_data.csv')

# # load data - Set 2
# amz_rev = pd.read_csv('/Users/pavansingh/Library/CloudStorage/GoogleDrive-pavansingho23@gmail.com/My Drive/Portfolio/Masters-Dissertation/Code/Data/set2_data.csv')

# # load data - Set 3
# amz_rev = pd.read_csv('/Users/pavansingh/Library/CloudStorage/GoogleDrive-pavansingho23@gmail.com/My Drive/Portfolio/Masters-Dissertation/Code/Data/set3_data.csv')

# load data - Set 4
# amz_rev = pd.read_csv('/Users/pavansingh/Library/CloudStorage/GoogleDrive-pavansingho23@gmail.com/My Drive/Portfolio/Masters-Dissertation/Code/Data/set4_data.csv')

#  drop column 'Unnamed: 0' or 'Unnamed: 0.1'
amz_rev = amz_rev.drop(columns=['Unnamed: 0.1'])
amz_rev = amz_rev.drop(columns=['Unnamed: 0'])

# initial data view
display(amz_rev.head(3))
print("Shape of the data: ", amz_rev.shape)

## <a id='toc2_1_'></a>[Removing Unnecessary Columns](#toc0_)

We look at removing columns that are not needed for the analysis.


# see columns
print("Columns in the data: \n", amz_rev.columns)

# remove certain columns
amz_rev.drop(['verified', 'style', 'reviewTime', 'vote', 'summary', 'image'], axis=1, inplace=True)

# see updated dataframe
display(amz_rev.head(3))

# sort and order the data
amz_rev.sort_values(by=['asin', 'overall'], ascending=[True, False], inplace=True)

# reorder columns
amz_rev = amz_rev[['reviewerID', 'reviewerName', 'unixReviewTime', 'asin', 'reviewText', 'category', 'overall']]

# see updated dataframe
display(amz_rev.head(3))

## <a id='toc2_2_'></a>[NaN/Missing Values in the Data](#toc0_)

We look at NaN values in the data. If any review (row) has missing values (NaN) in any column listed here:
    - (description, title, brand, price)
then we remove that review (row) from the dataset. 



# how many nulls in each column
amz_rev.isnull().sum()

# remove rows with null values
amz_rev = amz_rev.dropna()

# shape of the data
print("Shape of the data: ", amz_rev.shape)

# see if any missing values
print("\nNumber of missing values: ", amz_rev.isnull().sum().sum())

# show count columns have missing values (as a percentage)
print("\nPercentage of missing values in each column: \n", round(amz_rev.isnull().sum()/len(amz_rev)*100,2))

# show count of rows with missing data per category (as a percentage)
print("\nPercentage of missing values in each category: \n", round(amz_rev.groupby(['category']).apply(lambda x: x.isnull().sum()).sum(axis=1)/len(amz_rev)*100,2))

## <a id='toc2_3_'></a>[Duplicates in the Data](#toc0_)

We looking at duplicates in the data.

A duplicate is defined as a review (row) that has the same values across all columns. We remove duplicates from the dataset.

# see if any duplicates
print("Number of duplicates: ", amz_rev.duplicated().sum())

# see duplicates
amz_rev[amz_rev.duplicated(keep=False)].sort_values(by=['asin']).head(4)

# remove duplicates
amz_rev.drop_duplicates(inplace=True)

# see updated dataframe
display(amz_rev.head(3))

# shape of the data
print("Shape of the data: ", amz_rev.shape)

# # data set 1
amz_rev = amz_rev[amz_rev['asin'].isin(amz_rev.groupby('asin').size().reset_index(name='counts').query('counts >= 10')['asin'])]
amz_rev = amz_rev[amz_rev['reviewerID'].isin(amz_rev.groupby('reviewerID').size().reset_index(name='counts').query('counts >= 10')['reviewerID'])]
print(amz_rev.shape)

# # data set 2
# amz_rev = amz_rev[amz_rev['asin'].isin(amz_rev.groupby('asin').size().reset_index(name='counts').query('counts >= 12')['asin'])]
# amz_rev = amz_rev[amz_rev['reviewerID'].isin(amz_rev.groupby('reviewerID').size().reset_index(name='counts').query('counts >= 12')['reviewerID'])]
# print(amz_rev.shape)

# # data set 3
# amz_rev = amz_rev[amz_rev['asin'].isin(amz_rev.groupby('asin').size().reset_index(name='counts').query('counts >= 14')['asin'])]
# amz_rev = amz_rev[amz_rev['reviewerID'].isin(amz_rev.groupby('reviewerID').size().reset_index(name='counts').query('counts >= 14')['reviewerID'])]
# print(amz_rev.shape)

# data set 4
# amz_rev = amz_rev[amz_rev['asin'].isin(amz_rev.groupby('asin').size().reset_index(name='counts').query('counts >= 20')['asin'])]
# amz_rev = amz_rev[amz_rev['reviewerID'].isin(amz_rev.groupby('reviewerID').size().reset_index(name='counts').query('counts >= 20')['reviewerID'])]
# print(amz_rev.shape)

## <a id='toc2_5_'></a>[Standardizing and Checking Data Types](#toc0_)

We turn to standardizing and checking data types. 

# change unixReviewTime to datetime
amz_rev['unixReviewTime'] = pd.to_datetime(amz_rev['unixReviewTime'], unit='s')

# rename column: unixReviewTime to reviewTime
amz_rev.rename(columns={'unixReviewTime': 'reviewTime'}, inplace=True)

# see updated dataframe
display(amz_rev.head(3))

# check data types
amz_rev.dtypes

## <a id='toc2_6_'></a>[Normalizing Data](#toc0_)

We look at normalizing data (price, votes, etc. ) We used min-max. 

# get min and max for ratings
min_rating = amz_rev['overall'].min()
max_rating = amz_rev['overall'].max()

# Normalize the ratings to a range from 0 to 1
amz_rev['normalized_rating'] = (amz_rev['overall'] - min_rating) / (max_rating - min_rating)

# see updated dataframe
amz_rev.head(3)

# see summary of normalized ratings
amz_rev['normalized_rating'].describe()


The '`normalized_rating`' column will contain the normalized ratings between 0 and 1, where 0 corresponds to the minimum rating and 1 corresponds to the maximum rating. Note, the '`normalized_rating`' column is created by subtracting the minimum value from each rating and dividing it by the range (maximum value minus minimum value).

Ratings which were 1 are now 0 and ratings which were 5 are now 1 etc. 

# see updated dataframe
display(amz_rev.head(3))


# <a id='toc4_'></a>[Dealing with Text Data](#toc0_)

We deal with text data. We have the following columns that contain text data:
- description
- title
- brand
- category
- reviewText


We handle each of these columns separately. We look out for: 

**Identify special characters or symbols**: Look for any special characters, symbols, or non-alphanumeric characters that may need to be cleaned or removed. These characters can sometimes interfere with downstream analysis or modeling.

**Handle HTML tags or formatting**: If the 'description' column contains HTML tags or formatting, you may consider removing them or converting them into plain text

## <a id='toc4_1_'></a>[Text Normalization](#toc0_)

**Text Normalization**: Text normalization is the process of transforming text into a single canonical form that it might not have had before. This is done by removing unnecessary characters, such as punctuation or special characters; converting all letters to lowercase or uppercase; and/or expanding abbreviations.


### <a id='toc4_1_5_'></a>[ReviewText](#toc0_)

# see brand
amz_rev['reviewText']

# create function to clean brand column
def clean_brand(text):
    # Remove square brackets, single quotes, HTML tags, double quotes, commas, URLs, and CSS styling
    text = re.sub(r"\[|\]|\'|<[^>]*>|\"|,|\\https?://[^\s]+|{[^}]+}", "", text)
    
    # Remove extra whitespace
    text = " ".join(text.split())
    
    # Remove common update notes
    update_regex = r'update \d+/\d+/\d+:'
    text = re.sub(update_regex, '', text)

    # Remove trailing punctuation
    text = re.sub(r"\.$", "", text)

    # remove dates
    text = re.sub(r'\d{2}/\d{2}/\d{2}', '', text)

    
    # Convert to lowercase
    text = text.lower()
    
    return text


# apply function to brand column
amz_rev['reviewText'] = amz_rev['reviewText'].apply(clean_brand)

# see updated dataframe
amz_rev['reviewText']

# check any special characters in title
amz_rev[amz_rev['reviewText'].str.contains("[^a-zA-Z0-9\s]")]

# replace special characters in title
amz_rev['reviewText'] = amz_rev['reviewText'].str.replace("[^a-zA-Z0-9\s]", "")

# count of records with special character in the title
print("Number records with special character in reviewText:", amz_rev[amz_rev['reviewText'].str.contains("[^a-zA-Z0-9\s]")]['asin'].unique().size)

***
## <a id='toc4_2_'></a>[Text Cleaning](#toc0_)

We specifically explore: 
1. Tokenization
2. Stopword Removal
3. Stemming and Lemmatization
4. Handling Noise and Irrelevant Information

We do this for only the `reviewText` columns and the `description` columns. 


display(amz_rev.head(6))

### <a id='toc4_2_1_'></a>[Tokenization](#toc0_)

**Tokenization**: Tokenization is the process of splitting text into smaller chunks, called tokens. Tokens are usually words, sentences, or individual characters. Tokenization is a crucial step in text analysis, as the meaning of a word can fundamentally change based on the tokens surrounding it.


import nltk
from nltk.tokenize import word_tokenize
nltk.download('punkt')

# Assuming amz_rev.reviewText are Pandas Series
# Tokenize the text using a loop

tokens_reviewText = []
for text in amz_rev.reviewText:
    if isinstance(text, str):
        tokens_reviewText.append(word_tokenize(text))
    else:
        tokens_reviewText.append([])

# shape of tokens
print("Shape of tokens_reviewText:", len(tokens_reviewText))

# shape of reviewText
print("Shape of reviewText:", len(amz_rev.reviewText))

# Print the tokens
print("Some output of review text tokens:", tokens_reviewText[0:2])

### <a id='toc4_2_2_'></a>[ Stopword Removal](#toc0_)

**Stopword Removal**: Stopwords are words that are commonly used in the English language, such as "the," "a," "an," "is," and "are." These words are often removed from text during preprocessing as they can negatively impact downstream analysis, such as natural language processing and machine learning.


import nltk
from nltk.corpus import stopwords
from nltk.tokenize import word_tokenize
nltk.download('stopwords')

# Get the set of English stopwords
stopword_set = set(stopwords.words('english'))

# Remove stopwords from the tokens
filtered_tokens_reviewText = [[word for word in tokens if word.lower() not in stopword_set] for tokens in tokens_reviewText]


# Print the filtered tokens
print("Review Text tokens after stop word removal:", filtered_tokens_reviewText[0:2])

### <a id='toc4_2_3_'></a>[Stemming and Lemmatization](#toc0_)


**Stemming and Lemmatization**: Stemming is the process of reducing a word to its stem, or its root form. For example, "fishing," "fished," "fisher" all reduce to the stem "fish." Lemmatization is similar to stemming, but it brings context to the words. So, it links words with similar meaning to one word. For example, "better" and "good" are lemmatized to "good."

Stemming involves removing prefixes and suffixes from words to obtain the root form. For example, the word "running" would be stemmed to "run." Stemming is a simpler and faster process but may result in the root form not being an actual word. This can lead to potential loss of meaning or incorrect interpretations.

On the other hand, lemmatiation aims to determine the lemma or dictionary form of a word. It takes into account the word's context and part of speech, ensuring that the resulting lemma is a valid word. For example, the word "running" would be lemmatized to "run." Lemmatization provides more accurate results but can be computationally more expensive compared to stemming.

If preserving the exact meaning and interpretability of words is crucial for your recommender system, lemmatization would be a better choice. However, if speed and simplicity are more important, stemming could be sufficient.

from nltk.stem import PorterStemmer, WordNetLemmatizer
from nltk.tokenize import word_tokenize
nltk.download('punkt')

# Stemming
stemmer = PorterStemmer()
stemmed_words_revText = [[stemmer.stem(word) for word in tokens] for tokens in filtered_tokens_reviewText]

# Lemmatization
lemmatizer = WordNetLemmatizer()
lemmatized_words_revText = [[lemmatizer.lemmatize(word) for word in tokens] for tokens in filtered_tokens_reviewText]


print("Original words:", filtered_tokens_reviewText[0:2])
print("Stemmed words:", stemmed_words_revText[0:2])
print("Lemmatized words:", lemmatized_words_revText[0:2])

***
# <a id='toc6_'></a>[Creating New Dataset and Saving](#toc0_)

We create a new dataset using all the cleaning and processing done above and save it as a csv file called `data_clean.csv`.

# get shapes of all generated dataframes
print("Shape of stemmed_words_revText:", len(stemmed_words_revText))
print("Shape of lemmatized_words_revText:", len(lemmatized_words_revText))
print("Shape of filtered_tokens_reviewText:", len(filtered_tokens_reviewText))

# attach stemmed words to dataframe
amz_rev['stemmed_words_revText'] = stemmed_words_revText

# attach lemmitized words to dataframe
amz_rev['lemmatized_words_revText'] = lemmatized_words_revText

# attach filtered tokens to dataframe
amz_rev['filtered_tokens_revText'] = filtered_tokens_reviewText


# see resulting dataframe
amz_rev.head(6)

# # save cleaned data
amz_rev.to_csv("Data/set1_data_cleaned.csv", index=False)

# # save cleaned data
# amz_rev.to_csv("Data/set2_data_cleaned.csv", index=False)

# # save cleaned data
# amz_rev.to_csv("Data/set3_data_cleaned.csv", index=False)

# save cleaned data
# amz_rev.to_csv("Data/set4_data_cleaned.csv", index=False)


# stats of cleaned data
print("Shape of cleaned data:", amz_rev.shape)
print("\nNumber of unique products:", amz_rev['asin'].unique().size)
print("\nNumber of unique users:", amz_rev['reviewerID'].unique().size)
print("\nNumber of unique categories:", amz_rev['category'].unique().size)
print("\nNumber of reviews per category:\n", amz_rev['category'].value_counts())
print("\nColumns available:", amz_rev.columns)
